🚧 Update follow comment
